Break out by-name-closure creation from ElimByName
