+++ compiler/src/dotty/tools/dotc/transform/ElimByName.scala
@@ -20,27 +20,12 @@ import ast.Trees._
  *
  *  For the types of parameter symbols:
  *
- *      => T        ==>    () => T
+ *         => T       ==>    () => T
  *
- *  Note that `=> T` types are not eliminated in MethodTypes. This is done later at erasure.
- *  Terms are rewritten as follows:
+ *  For cbn parameter values
  *
- *      x           ==>    x.apply()   if x is a parameter that had type => T
- *
- *  Arguments to call-by-name parameters are translated as follows. First, the argument is
- *  rewritten by the rules
- *
- *      e.apply()   ==>    e           if e.apply() is an argument to a call-by-name parameter
- *      expr        ==>    () => expr  if other expr is an argument to a call-by-name parameter
- *
- *  This makes the argument compatible with a parameter type of () => T, which will be the
- *  formal parameter type at erasure. But to be -Ycheckable until then, any argument
- *  ARG rewritten by the rules above is again wrapped in an application DummyApply(ARG)
- *  where
- *
- *     DummyApply: [T](() => T): T
- *
- *  is a synthetic method defined in Definitions. Erasure will later strip these DummyApply wrappers.
+ *         x          ==>    x()
+ *         CbnArg(x)  ==>    DummyApply(x)
  *
  *  Note: This scheme to have inconsistent types between method types (whose formal types are still
  *  ExprTypes and parameter valdefs (which are now FunctionTypes) is not pretty. There are two
@@ -53,7 +38,7 @@ import ast.Trees._
  *  Option 2: Merge ElimByName with erasure, or have it run immediately before. This has not been
  *  tried yet.
  */
-class ElimByName extends MiniPhaseTransform with InfoTransformer { thisTransformer =>
+class ElimByName extends ByNameClosures with InfoTransformer { thisTransformer =>
   import ast.tpd._
 
   override def phaseName: String = "elimByName"
@@ -62,52 +47,10 @@ class ElimByName extends MiniPhaseTransform with InfoTransformer { thisTransform
     // assumes idents and selects have symbols; interferes with splitter distribution
     // that's why it's "after group".
 
-  /** The info of the tree's symbol at phase Nullarify (i.e. before transformation) */
-  private def originalDenotation(tree: Tree)(implicit ctx: Context) =
-    tree.symbol.denot(ctx.withPhase(thisTransformer))
-
-  override def transformApply(tree: Apply)(implicit ctx: Context, info: TransformerInfo): Tree =
-    ctx.traceIndented(s"transforming ${tree.show} at phase ${ctx.phase}", show = true) {
-
-    def transformArg(arg: Tree, formal: Type): Tree = formal.dealias match {
-      case formalExpr: ExprType =>
-        var argType = arg.tpe.widenIfUnstable
-        if (defn.isBottomType(argType)) argType = formal.widenExpr
-        val argFun = arg match {
-          case Apply(Select(qual, nme.apply), Nil)
-          if qual.tpe.derivesFrom(defn.FunctionClass(0)) && isPureExpr(qual) =>
-            qual
-          case _ =>
-            val inSuper = if (ctx.mode.is(Mode.InSuperCall)) InSuperCall else EmptyFlags
-            val meth = ctx.newSymbol(
-                ctx.owner, nme.ANON_FUN, Synthetic | Method | inSuper, MethodType(Nil, Nil, argType))
-            Closure(meth, _ =>
-              atGroupEnd { implicit ctx: Context =>
-                arg.changeOwner(ctx.owner, meth)
-              }
-            )
-        }
-        ref(defn.dummyApply).appliedToType(argType).appliedTo(argFun)
-      case _ =>
-        arg
-    }
-
-    val mt @ MethodType(_) = tree.fun.tpe.widen
-    val args1 = tree.args.zipWithConserve(mt.paramInfos)(transformArg)
-    cpy.Apply(tree)(tree.fun, args1)
-  }
-
-  /** If denotation had an ExprType before, it now gets a function type */
-  private def exprBecomesFunction(symd: SymDenotation)(implicit ctx: Context) =
-    (symd is Param) || (symd is (ParamAccessor, butNot = Method))
-
   /** Map `tree` to `tree.apply()` is `ftree` was of ExprType and becomes now a function */
-  private def applyIfFunction(tree: Tree, ftree: Tree)(implicit ctx: Context) = {
-    val origDenot = originalDenotation(ftree)
-    if (exprBecomesFunction(origDenot) && (origDenot.info.isInstanceOf[ExprType]))
-      tree.select(defn.Function0_apply).appliedToNone
+  private def applyIfFunction(tree: Tree, ftree: Tree)(implicit ctx: Context) =
+    if (isByNameRef(ftree)) tree.select(defn.Function0_apply).appliedToNone
     else tree
-  }
 
   override def transformIdent(tree: Ident)(implicit ctx: Context, info: TransformerInfo): Tree =
     applyIfFunction(tree, tree)
